Правки наставника: Убраны пропсы из компонента cities-list; Переделана константа CITIES; Реструктуризирована папка hooks; Логика компонента main-page вынесена в хук current-offers.
