+++ src/hooks/use-leaflet-map.tsx
@@ -2,7 +2,10 @@ import { useEffect, useState, useRef, MutableRefObject } from 'react';
 import { Map as LeafletMap, TileLayer } from 'leaflet';
 import { City } from '../types/offer';
 
-function useLeafletMap(mapRef: MutableRefObject<HTMLElement | null>, city: City): LeafletMap | null {
+export function useLeafletMap(
+	mapRef: MutableRefObject<HTMLElement | null>,
+	city: City
+): LeafletMap | null {
 	const [leafletMap, setLeafletMap] = useState<LeafletMap | null>(null);
 	const isRenderedRef = useRef<boolean>(false);
 
@@ -11,16 +14,16 @@ function useLeafletMap(mapRef: MutableRefObject<HTMLElement | null>, city: City)
 			const instance = new LeafletMap(mapRef.current, {
 				center: {
 					lat: city.location.latitude,
-					lng: city.location.longitude
+					lng: city.location.longitude,
 				},
-				zoom: city.location.zoom
+				zoom: city.location.zoom,
 			});
 
 			const layer = new TileLayer(
 				'https://{s}.basemaps.cartocdn.com/rastertiles/voyager/{z}/{x}/{y}{r}.png',
 				{
 					attribution:
-						'&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors &copy; <a href="https://carto.com/attributions">CARTO</a>'
+						'&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors &copy; <a href="https://carto.com/attributions">CARTO</a>',
 				}
 			);
 
@@ -33,5 +36,3 @@ function useLeafletMap(mapRef: MutableRefObject<HTMLElement | null>, city: City)
 
 	return leafletMap;
 }
-
-export default useLeafletMap;
